Revert "Feature/lib serverproto"
