+++ .gitmodules
@@ -7,9 +7,3 @@
 [submodule "src/depends/mongo-cxx-driver"]
 	path = src/depends/mongo-cxx-driver
 	url = https://github.com/mongodb/mongo-cxx-driver.git
-[submodule "src/depends/grpc"]
-	path = src/depends/grpc
-	url = https://github.com/grpc/grpc
-[submodule "src/depends/protobuf"]
-	path = src/depends/protobuf
-	url = https://github.com/protocolbuffers/protobuf
