+++ docker/Dockerfile
@@ -25,9 +25,6 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     libprotobuf-dev \
     protobuf-compiler \
     libcurl4-openssl-dev \
-    libtool \
-    autoconf \
-    golang \
     python python-pip \
     && rm -rf /var/lib/apt/lists/* \
     && pip install netaddr
