+++ src/libProtoServer/ServerMessages.proto
@@ -2,6 +2,8 @@ syntax = "proto2";
 
 package ZilliqaMessage;
 
+import "ZilliqaMessage.proto";
+
 
 message ProtoTxId
 {
@@ -27,6 +29,11 @@ message ProtoAddress
 message ProtoBlockNum
 {
     required uint64 blocknum = 1;
+}
+
+message ProtoStorageInfo
+{
+
 }
 
 message ProtoShardingStruct
